Refactor CourseView folder structure, modularize sockets
