--- client/components/ClassView/CourseContainer.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import DoughnutChart from './DoughnutChart.jsx';
-import CourseInfo from './CourseInfo.jsx';
-
-const CourseContainer = () => (
-  <div className="container-fluid">
-    <div>
-        <h4>Course Health</h4>
-        <hr />
-    </div>
-    <CourseInfo />
-  </div>
-);
-
-export default CourseContainer;
