+++ client/components/CourseView/AssignmentsTab/Assignment.jsx
@@ -1,47 +1,45 @@
 import React, { Component, PropTypes } from 'react';
-import actions from '../../actions/index.js';
+import actions from '../../../actions/index.js';
 import { connect } from 'react-redux';
 
-class StudentTab extends Component {
+class Assignment extends Component {
   // when a single student is clicked
   // get the current student and display their info
   render() {
     return (
-      <div>
+      <div className="row">
         <button
           type="button"
           className="btn btn-secondary back"
-          onClick={ this.props.handleBackButton }
+          onClick={ () => this.props.handleBackButton() }
         >
           Back
         </button>
-        {this.props.displayedStudent.name}
+        {this.props.displayedAssignment.name}
       </div>
     );
   }
 }
 
 const mapStateToProps = (state) => (
-  {
-    displayedStudent: state.displayedStudent,
-  }
+  { displayedAssignment: state.displayedAssignment }
 );
 
 // go back to all students when back is pressed
 const mapDispatchToProps = (dispatch) => (
   {
     handleBackButton: () => {
-      dispatch(actions.switchTabs('Students'));
+      dispatch(actions.switchTabs('Assignments'));
     },
   }
 );
 
-StudentTab.propTypes = {
-  displayedStudent: PropTypes.object,
+Assignment.propTypes = {
+  displayedAssignment: PropTypes.object,
   handleBackButton: PropTypes.func,
 };
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(StudentTab);
+)(Assignment);
