+++ client/components/CourseView/CourseView.jsx
@@ -1,13 +1,12 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import actions from '../../actions/index.js';
-import AllStudentsTab from './AllStudentsTab.jsx';
-import AssignmentsTab from './AssignmentsTab.jsx';
-import Student from './Student.jsx';
-import Assignment from './Assignment.jsx';
-import DoughnutChart from './DoughnutChart.jsx';
-import BarChart from './BarChart.jsx';
-
+import AllStudentsTab from './StudentsTab/AllStudentsTab.jsx';
+import AssignmentsTab from './AssignmentsTab/AssignmentsTab.jsx';
+import StudentCard from './StudentsTab/StudentCard.jsx';
+import Assignment from './AssignmentsTab/Assignment.jsx';
+import DoughnutChart from './Charts/DoughnutChart.jsx';
+import BarChart from './Charts/BarChart.jsx';
 
 class CourseInfo extends Component {
   // depending on which tab is clicked, render Assignments, AllStudents, or Student component
@@ -17,7 +16,7 @@ class CourseInfo extends Component {
         <AllStudentsTab />
       );
     } else if (this.props.tabView === 'Student') {
-      return <Student />;
+      return <StudentCard />;
     } else if (this.props.tabView === 'Assignment') {
       return <Assignment />;
     }
@@ -51,22 +50,26 @@ class CourseInfo extends Component {
 
   render() {
     return (
-      <div className="course-info">
-        <div>
-          {this.handleCharts()}
-        </div>
-        <ul className="course-info-nav nav nav-tabs">
-          <li
-            role="presentation"
-            className="course-info-nav-tab assignments-tab"
-            onClick={ () => this.props.handleTab('Assignments') }
-          >
-            <a>View Assignments</a>
-          </li>
-          { this.displayStudentsTab() }
-        </ul>
-        <div className="course-info-content">
-          { this.handleTabs() }
+      <div className="container-fluid">
+        <h4>Course Health</h4>
+        <hr />
+        <div className="course-info">
+          <div>
+            {this.handleCharts()}
+          </div>
+          <ul className="course-info-nav nav nav-tabs">
+            <li
+              role="presentation"
+              className="course-info-nav-tab assignments-tab"
+              onClick={ () => this.props.handleTab('Assignments') }
+              >
+              <a>View Assignments</a>
+            </li>
+            { this.displayStudentsTab() }
+          </ul>
+          <div className="course-info-content">
+            { this.handleTabs() }
+          </div>
         </div>
       </div>
     );
