+++ client/components/CourseView/StudentsTab/StudentCard.jsx
@@ -1,45 +1,47 @@
 import React, { Component, PropTypes } from 'react';
-import actions from '../../actions/index.js';
+import actions from '../../../actions/index.js';
 import { connect } from 'react-redux';
 
-class Assignment extends Component {
+class StudentTab extends Component {
   // when a single student is clicked
   // get the current student and display their info
   render() {
     return (
-      <div className="row">
+      <div>
         <button
           type="button"
           className="btn btn-secondary back"
-          onClick={ () => this.props.handleBackButton() }
+          onClick={ this.props.handleBackButton }
         >
           Back
         </button>
-        {this.props.displayedAssignment.name}
+        {this.props.displayedStudent.name}
       </div>
     );
   }
 }
 
 const mapStateToProps = (state) => (
-  { displayedAssignment: state.displayedAssignment }
+  {
+    displayedStudent: state.displayedStudent,
+  }
 );
 
 // go back to all students when back is pressed
 const mapDispatchToProps = (dispatch) => (
   {
     handleBackButton: () => {
-      dispatch(actions.switchTabs('Assignments'));
+      dispatch(actions.switchTabs('Students'));
     },
   }
 );
 
-Assignment.propTypes = {
-  displayedAssignment: PropTypes.object,
+StudentTab.propTypes = {
+  displayedStudent: PropTypes.object,
   handleBackButton: PropTypes.func,
 };
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(Assignment);
+)(StudentTab);
