+++ client/index.html
@@ -35,8 +35,5 @@
     <div id="app"></div>
     <script src="./dist/bundle.js" async></script>
     <script src="/socket.io/socket.io.js"></script>
-    <script>
-      var socket = io();
-    </script>
   </body>
 </html>
