+++ client/utils/socket.js
@@ -1,0 +1,1 @@
+export default io();
