+++ tests/client/classViewTests/classContainer.tests.js
@@ -2,13 +2,13 @@ import React from 'react';
 import TestUtils from 'react-addons-test-utils'; // Alternately could use the DOM API
 import configureStore from '../../../client/store/configureStore';
 import { Provider } from 'react-redux';
-import CourseContainer from '../../../client/components/ClassView/CourseContainer.jsx';
+import CourseView from '../../../client/components/CourseView/CourseView.jsx';
 const store = configureStore();
 
-xdescribe('CourseContainer', () => {
+xdescribe('CourseView', () => {
   const courseContainer = TestUtils.renderIntoDocument(
     <Provider store={store}>
-      <CourseContainer />
+      <CourseView />
     </Provider>
   );
